requirements.txt
streamlit
google-generativeai
pytest
pandas
numpy
python-dotenv
tqdm